requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
tkinter # Usually comes with Python installation
